Remove Node class
