+++ distributed/node.py
@@ -1,53 +1,20 @@
-import asyncio
 import logging
 import warnings
 import weakref
 
-from tornado import gen
-from tornado.ioloop import IOLoop
 from tornado.httpserver import HTTPServer
 import tlz
 import dask
 
 from .comm import get_tcp_server_address
 from .comm import get_address_host
-from .core import Server, ConnectionPool
+from .core import Server
 from .http.routing import RoutingApplication
 from .versions import get_versions
-from .utils import DequeHandler, TimeoutError, clean_dashboard_address, ignoring
+from .utils import DequeHandler, clean_dashboard_address, ignoring
 
 
-class Node:
-    """
-    Base class for nodes in a distributed cluster.
-    """
-
-    def __init__(
-        self,
-        connection_limit=512,
-        deserialize=True,
-        connection_args=None,
-        io_loop=None,
-        serializers=None,
-        deserializers=None,
-        timeout=None,
-    ):
-        self.io_loop = io_loop or IOLoop.current()
-        self.rpc = ConnectionPool(
-            limit=connection_limit,
-            deserialize=deserialize,
-            serializers=serializers,
-            deserializers=deserializers,
-            connection_args=connection_args,
-            timeout=timeout,
-            server=self,
-        )
-
-    async def start(self):
-        await self.rpc.start()
-
-
-class ServerNode(Node, Server):
+class ServerNode(Server):
     """
     Base class for server nodes in a distributed cluster.
     """
@@ -56,40 +23,6 @@ class ServerNode(Node, Server):
 
     # XXX avoid inheriting from Server? there is some large potential for confusion
     # between base and derived attribute namespaces...
-
-    def __init__(
-        self,
-        handlers=None,
-        blocked_handlers=None,
-        stream_handlers=None,
-        connection_limit=512,
-        deserialize=True,
-        connection_args=None,
-        io_loop=None,
-        serializers=None,
-        deserializers=None,
-        timeout=None,
-    ):
-        Node.__init__(
-            self,
-            deserialize=deserialize,
-            connection_limit=connection_limit,
-            connection_args=connection_args,
-            io_loop=io_loop,
-            serializers=serializers,
-            deserializers=deserializers,
-            timeout=timeout,
-        )
-        Server.__init__(
-            self,
-            handlers=handlers,
-            blocked_handlers=blocked_handlers,
-            stream_handlers=stream_handlers,
-            connection_limit=connection_limit,
-            deserialize=deserialize,
-            io_loop=self.io_loop,
-        )
-        self._startup_lock = asyncio.Lock()
 
     def versions(self, comm=None, packages=None):
         return get_versions(packages=packages)
@@ -161,42 +94,6 @@ class ServerNode(Node, Server):
             L = deque_handler.deque
             L = [L[-i] for i in range(min(n, len(L)))]
         return [(msg.levelname, deque_handler.format(msg)) for msg in L]
-
-    async def __aenter__(self):
-        await self
-        return self
-
-    async def __aexit__(self, typ, value, traceback):
-        await self.close()
-
-    def __await__(self):
-        async def _():
-            timeout = getattr(self, "death_timeout", 0)
-            async with self._startup_lock:
-                if self.status == "running":
-                    return self
-                if timeout:
-                    try:
-                        await asyncio.wait_for(self.start(), timeout=timeout)
-                        self.status = "running"
-                    except Exception:
-                        await self.close(timeout=1)
-                        raise TimeoutError(
-                            "{} failed to start in {} seconds".format(
-                                type(self).__name__, timeout
-                            )
-                        )
-                else:
-                    await self.start()
-                    self.status = "running"
-            return self
-
-        return _().__await__()
-
-    async def start(self):
-        # subclasses should implement their own start method whichs calls super().start()
-        await Node.start(self)
-        return self
 
     def start_http_server(
         self, routes, dashboard_address, default_port=0, ssl_options=None,
